api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pandas>=1.3.0
numpy>=1.21.0
mlflow>=3.0.0
lightgbm>=3.3.5
scikit-learn>=1.0.0
pydantic>=2.0.0
evidently>=0.4.0
prometheus-client>=0.17.0
python-dotenv>=1.0.0
